Code on a single component for greater legibility
